+++ ricaun.Revit.UI.Tests/Base/BaseImage.cs
@@ -1,0 +1,7 @@
+﻿namespace ricaun.Revit.UI.Tests
+{
+    public class BaseImage
+    {
+        public const string Revit = "/UIFrameworkRes;component/ribbon/images/revit.ico";
+    }
+}