+++ ricaun.Revit.UI/RibbonItemExtension.cs
@@ -149,7 +149,7 @@ namespace ricaun.Revit.UI
 
         #region Set RibbonButton
         /// <summary>
-        /// Set RibbonButton/ComboBox Image
+        /// Set RibbonButton/ComboBox/TextBox Image
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
@@ -166,12 +166,13 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton/ComboBox LargeImage
+        /// Set RibbonButton/ComboBox/TextBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem"></typeparam>
         /// <param name="ribbonItem"></param>
         /// <param name="largeImage"></param>
         /// <returns></returns>
+        /// <remarks>When <see cref="ComboBox"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, string largeImage) where TRibbonItem : RibbonItem
         {
             var bitmapSource = largeImage?.GetBitmapSource();
@@ -204,12 +205,13 @@ namespace ricaun.Revit.UI
         }
 
         /// <summary>
-        /// Set RibbonButton/ComboBox LargeImage
+        /// Set RibbonButton/ComboBox/TextBox LargeImage
         /// </summary>
         /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
         /// <param name="ribbonItem"></param>
         /// <param name="largeImage"></param>
         /// <returns></returns>
+        /// <remarks>When <see cref="ComboBox"/> or <see cref="TextBox"/> does not have LargeImage, the Image is changed instead.</remarks>
         public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItem
         {
             if (ribbonItem is RibbonButton ribbonButton)
@@ -221,6 +223,9 @@ namespace ricaun.Revit.UI
 
             else if (ribbonItem is ComboBox comboBox)
                 comboBox.SetImage(largeImage);
+
+            else if (ribbonItem is TextBox textBox)
+                textBox.SetImage(largeImage);
 
             return ribbonItem;
         }