+++ ricaun.Revit.UI/RibbonStackExtension.cs
@@ -14,10 +14,22 @@ namespace ricaun.Revit.UI
         /// <param name="ribbonPanel"></param>
         /// <param name="ribbonItems"></param>
         /// <returns></returns>
-        /// <remarks>The <paramref name="ribbonItems"/> is divided into groups of a maximum of 3 in each RibbonRowPanel</remarks>
+        /// <remarks>The <paramref name="ribbonItems"/> is divided into groups of a maximum of 3 in each RibbonRowPanel with Image size.</remarks>
         public static Autodesk.Windows.RibbonRowPanel[] RowStackedItems(this RibbonPanel ribbonPanel, params RibbonItem[] ribbonItems)
         {
             return ribbonPanel.CreateRowStackedItemsWithMax(ribbonItems);
+        }
+
+        /// <summary>
+        /// Create Row Panels and move the <paramref name="ribbonItems"/> to the new panels
+        /// </summary>
+        /// <param name="ribbonPanel"></param>
+        /// <param name="ribbonItems"></param>
+        /// <returns></returns>
+        /// <remarks>The <paramref name="ribbonItems"/> is divided into groups of a maximum of 2 in each RibbonRowPanel with LargeImage size.</remarks>
+        public static Autodesk.Windows.RibbonRowPanel[] RowLargeStackedItems(this RibbonPanel ribbonPanel, params RibbonItem[] ribbonItems)
+        {
+            return ribbonPanel.CreateRowStackedItemsWithMax(2, ribbonItems).SetRibbonItemSize();
         }
 
         /// <summary>