+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.7] / 2023-11-17
+### Features
+- LargeImage changes `TextBox` Image
+- Create `RowLargeStackedItems` with large image and two itens
+### Updated
+- Update Example `Icons` to get random icon using `UIFrameworkRes`
+### Tests
+- Test `ComboBox` and `TextBox` Image/LargeImage
+- Test `RowLargeStackedItems` with large image and two itens
+
 ## [0.5.6] / 2023-11-08
 ### Features
 - Update Remove and QuickAccess
@@ -291,6 +301,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.5.7]: ../../compare/0.5.6...0.5.7
 [0.5.6]: ../../compare/0.5.5...0.5.6
 [0.5.5]: ../../compare/0.5.4...0.5.5
 [0.5.4]: ../../compare/0.5.3...0.5.4